Refactor : line 관련 control 분리
<description>
line의 데이터를 다루는 컨트롤러를 line-control.js파일에 넣었다.
